requirements.txt
python-dotenv
requests
deepgram-sdk
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
